requirements.txt
dlt[filesystem,postgres,dbt]>=1.10.0
python-dotenv>=0.21.0
dbt-core>=1.9.0
dbt-postgres>=1.9.0
PyYAML>=6.0
psycopg2-binary>=2.9.0
